examples/text_content_manipulation/requirements.txt
tensorflow-gpu
numpy
munkres
nltk
h5py
pyxdameraulevenshtein
scikit-learn
